app/requirement.txt
Flask==2.2.3
Werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.28.2
